refactor: move properties to configuration
